% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.data.frame.Experiment}
\alias{as.data.frame.Experiment}
\alias{as.data.frame.SingleEndSamples}
\alias{as.data.frame.PairedSamples}
\title{Convert an experiment into a data frame}
\usage{
\method{as.data.frame}{Experiment}(x, row.names = NULL,
  optional = FALSE, ...)

\method{as.data.frame}{SingleEndSamples}(x, row.names = NULL,
  optional = FALSE, ...)

\method{as.data.frame}{PairedSamples}(x, row.names = NULL,
  optional = FALSE, ...)
}
\arguments{
\item{x}{\link{Experiment}}

\item{row.names}{ignored}

\item{optional}{ignored}

\item{...}{ignored}
}
\value{
a data frame
}
\description{
Convert an experiment into a data frame
}
