% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{byAll}
\alias{byAll}
\title{Run a function on the whole experiment}
\usage{
byAll(obj, fun, ..., thread = 1)
}
\arguments{
\item{obj}{the experiment}

\item{fun}{the function}

\item{...}{additional parameters to fun}

\item{thread}{the number of parallel process to execute}
}
\value{
an object
}
\description{
This function will execute a function call
on the whole experiment.
If threads > 1 and the doParallel package
has been installed, multiple processes will
be executed at the same time.
}
