% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{byMethods}
\alias{byMethods}
\alias{bySample}
\alias{byRun}
\alias{byFct}
\alias{byAll}
\alias{byFile}
\title{Set of methods to apply a function
over an Experiment object}
\usage{
bySample(obj, fun, ..., thread = 1)

byRun(obj, fun, ..., thread = 1)

byFct(obj, fun, by, ..., thread = 1)

byAll(obj, fun, ..., thread = 1)

byFile(obj, fun, ..., thread = 1)
}
\arguments{
\item{obj}{\link{Experiment}}

\item{fun}{function, the function that will be applied
to each subset of the \link{Experiment}}

\item{...}{additional parameters to fun}

\item{thread}{integer, the number of parallel process
to execute}

\item{by}{a factor. The experiment will be spitted according
to this factor}
}
\value{
a list. The length of the list depends on the
factor used to split the experiment. \code{byAll} returns
a single object.
}
\description{
If threads > 1 and the \link{doMC} package
has been installed, multiple processes will
be executed at the same time.

\code{bySample} the function is called splitting
the experiment into sample/s.

\code{byRun} the function is called splitting
the experiment into run/s.

\code{byFct} the function is called splitting
the experiment according to the factor provided.

\code{byAll} the function is called using all files
as input/output.

\code{byFile} the function is called on each
file separately.
}
