% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{bySample}
\alias{bySample}
\title{Run a function for each sample}
\usage{
bySample(obj, fun, ..., thread = 1)
}
\arguments{
\item{obj}{the experiment}

\item{fun}{the function}

\item{...}{additional parameters to fun}

\item{thread}{the number of parallel process to execute}
}
\value{
a list of object
}
\description{
This function will execute a function call
for each sample of the experiment provided.
If threads > 1 and the doParallel package
has been installed, multiple processes will
be executed at the same time.
}
