% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{changeExp}
\alias{changeExp}
\title{Change the experimen withing a \link{Task} object}
\usage{
changeExp(obj, new, ...)
}
\arguments{
\item{obj}{a \link{Task} object}

\item{new}{either a new \link{Experiment}
or a function that will be used to process
the \link{Experiment} contained in the Task}

\item{...}{additional argument that will be
passed to \code{new} if it is function
(ignored otherwise).}
}
\value{
a \link{Task} object
}
\description{
The experiment contained in a \link{Task} object
will be changed according to the \code{new}
parameter.
}
