% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{clearOutput}
\alias{clearOutput}
\title{Removal of output file/s from a \link{Experiment}
object}
\usage{
clearOutput(obj)
}
\arguments{
\item{obj}{a \link{Experiment}}
}
\value{
\code{TRUE} for each file correctly
removed
}
\description{
This method will remove any output file/s
stored in the given in the \link{Experiment}
object
}
