% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.evalCommandList}
\alias{.evalCommandList}
\title{Mothod used internally to eval
a system call}
\usage{
.evalCommandList(list, command, args, stdout, stderr, thread)
}
\arguments{
\item{list}{a list of environments (data frame)}

\item{command}{the command to be executed}

\item{args}{a name object. Thsi will be evalueted
and passed to the args param of \link{system2} function}

\item{stdout}{passed to stdout of \link{system2}}

\item{stderr}{passed to stderr of \link{system2}}

\item{thread}{numeric. The number of thread}
}
\value{
a list of bject returned by \link{system2} function
}
\description{
Mothod used internally to eval
a system call
}
