% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.evalFunBy}
\alias{.evalFunBy}
\title{Mothod used internally to eval
function}
\usage{
.evalFunBy(obj, fct, f, thread)
}
\arguments{
\item{obj}{a \link{Experiment} object}

\item{fct}{a factor}

\item{f}{a name object. Function parameters}

\item{thread}{numeric. The number of thread}
}
\value{
a list of bject returned by the function
}
\description{
Mothod used internally to eval
function
}
