% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.getExperiment}
\alias{.getExperiment}
\title{Constructor for \link{Experiment}.}
\usage{
.getExperiment(samples, run, step)
}
\arguments{
\item{samples}{character vector. Sample names}

\item{run}{character vector. Run ids}

\item{step}{character. The name of the analysis step}
}
\value{
an instance of \link{Experiment}
}
\description{
This functiona is used implicitly by other constructors
and should not be used directly.
}
