% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.getRunFromHeader}
\alias{.getRunFromHeader}
\title{Get run id from sequence header}
\usage{
.getRunFromHeader(path, nrecords = 1)
}
\arguments{
\item{path}{character vector. The paths to the sequence file/s}

\item{nrecords}{numeric. The number of record that must be read
before returning the run id}
}
\value{
the run id as character
}
\description{
This function is used internally by other
constructors to detect the run id of each file.
The id is parsed from the sequence header.
}
