% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Experiment-class}
\alias{Experiment-class}
\alias{Experiment}
\title{Main targeted experiment class}
\value{
a Targeted experiment
}
\description{
This class represents a sequencing experiment.
Each sample may have one or two files (single-end
or paired-end sequencing) and a run identifier.
Users should not inizialize this class unless their
are sure of what they are doing. Instances of this
class can be constructed using the method
}
\section{Slots}{

\describe{
\item{\code{samples}}{Character vector, sample names}

\item{\code{run}}{Character vector, the run id for each sample}

\item{\code{step}}{Character, the step of analysis}

\item{\code{n}}{Numeric vector, the number of samples}
}}

