% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{expFromFiles}
\alias{expFromFiles}
\title{Construct an Experiment from file names}
\usage{
expFromFiles(maindir = ".", pattern = ".fastq", samples = NULL,
  run = NULL, output = NULL, recursive = F)
}
\arguments{
\item{maindir}{Character, the path to the main directory}

\item{pattern}{Character or Character vector of length two.
If a single character, a single-end experiment will be created
otherwise the first element will be used to detected forward
sequence files and the second element will be used to detect
reverse sequence files.}

\item{samples}{either \code{NULL}, a Character vector, or a
function. If \code{NULL}, samples will have the same name
as the input files. If a vector, it will be used as vector
name for samples. If a function, input files will be converted
to sample names using the function provided.}

\item{run}{either \code{NULL}, a Character vector, or a
function. In the first case samples will hove no run assigned,
whereas in the other two the behaviour will be the same of the
\code{sample} option.}

\item{output}{either \code{NULL}, a Character vector, or a
function. Same behaviour as the \code{run} parameter.}

\item{recursive}{logical, should the listing recurse into directories?}
}
\value{
an \link{Experiment} object
}
\description{
This methods constructs an \link{Experiment} object
from file names in a given directory.
}
\seealso{
\code{\link[=formatSample]{formatSample()}}, \code{\link[=formatOutptut]{formatOutptut()}}, and \code{\link[=runFromHeader]{runFromHeader()}}
or \code{\link[=runFromBasedir]{runFromBasedir()}} for sample, output, and run formatting.
}
