% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatOutput}
\alias{formatOutput}
\title{Format output function}
\usage{
formatOutput(prefix, outdir = NULL)
}
\arguments{
\item{prefix}{character. The prefix to add}

\item{outdir}{character. The output directory}
}
\value{
a file path
}
\description{
A prefix will be added to the name fo the
file provided and its directory will be changed
as specified in the \link{outdir} param
}
