% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatSample}
\alias{formatSample}
\title{Format function for sample definition}
\usage{
formatSample(index = 1, sep = "[[:punct:]]")
}
\arguments{
\item{index}{numeric. The element to return}

\item{sep}{ragular expression. Input files will be splitted
accordingly}
}
\value{
formatted sample names
}
\description{
Sample names will be converted from file
names using this function. File names
will be splitted depending on the \link{sep}
param and the element corresponding to the \link{index}
will be returned
}
