% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fromOut}
\alias{fromOut}
\title{Format function to be used with
\link{runTask} method. This
function is a wrapper for the
\link{getExperimentFromOutput} function}
\usage{
fromOut(new.output)
}
\arguments{
\item{new.output}{either a character vector (or a matrix with two
columns for paired experiments), NULL, or a function.
If a character vector is provided, output file will be named
accordingly otherwise the specified function will be used
to convert input files into output ones. If NULL, the
resulting experiment will hove no output.}
}
\value{
a TargetedExperiment
}
\description{
Format function to be used with
\link{runTask} method. This
function is a wrapper for the
\link{getExperimentFromOutput} function
}
