% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identity}
\alias{identity}
\title{Format function to be used with
\link{runTask} method. This
function will return an experiment
object identitcal to the one
processed}
\usage{
identity()
}
\value{
the same experiment in teh input
}
\description{
Format function to be used with
\link{runTask} method. This
function will return an experiment
object identitcal to the one
processed
}
