% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{newTask}
\alias{newTask}
\title{Task constructor}
\usage{
newTask(exp)
}
\arguments{
\item{exp}{\link{Experiment}, the experiment on which the
task will be executed}
}
\value{
a \link{Task} object
}
\description{
It creates a \link{Task} object given
the experiment that will be
used as input.
}
