% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{out2exp}
\alias{out2exp}
\title{Get an experiment from the output of another one.}
\usage{
out2exp(obj, new.output = NULL)
}
\arguments{
\item{obj}{a Experiment. The experiment to convert}

\item{new.output}{either a character vector (or a matrix with two
columns for paired experiments), NULL, or a function.
If a character vector is provided, output file will be named
accordingly otherwise the specified function will be used
to convert input files into output ones. If NULL, the
resulting experiment will hove no output.}
}
\value{
a Experiment
}
\description{
This function will convert an experiment to another one
by inverting the output files as input. The new output files
will be generate depending on the \link{new.output} param
}
