% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{PairedSamples-class}
\alias{PairedSamples-class}
\alias{.PairedSamples}
\title{Paired sample class}
\value{
a paired experiment representation
}
\description{
This class represent a standard paired-end experiment.
In addition to the \link[=.TargetedExperiment]{TargetedExperiment}
class it provides a vector of forward and reverse file paths
togheter with a list of forward and reverse output
paths. All slots can be used in the context
of other methods such as \code{\link[=by_sample]{by_sample()}}, for
sample wise computation and \code{\link[=by_run]{by_run()}} for run wise
computation.
}
\details{
Instances of this class should be created using the
\code{\link[=getPairedFromFilename]{getPairedFromFilename()}} function.
}
\section{Slots}{

\describe{
\item{\code{forward}}{character vector. Forward file paths}

\item{\code{reverse}}{character vector. Reverse file paths}

\item{\code{forward.out}}{character vector. Forward output files}

\item{\code{reverse.out}}{character vector. Reverse output files}
}}

