% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_clean_RNA_sample_info_and_protein_expressions.R
\name{prepare_clean_RNA_sample_info_and_protein_expressions}
\alias{prepare_clean_RNA_sample_info_and_protein_expressions}
\title{prepare_clean_RNA_sample_info_and_protein_expressions}
\usage{
prepare_clean_RNA_sample_info_and_protein_expressions(
  RNA_sample_info,
  sample_id_col = c("Sample_ID", "Tumor_Sample_Barcode", "sample_id", "id"),
  expressions,
  gene_id_col = c("rowname", "gene_id", "Gene_ID", "ID"),
  mark_library_size = T,
  adjusted_by_library_size = T,
  tolerant_library_size_factor = 1.1,
  protein_coding_ensemble2symbol_table,
  ens_id_col = c("Gene_ID", "gene_id", "id"),
  symbol_col = c("Gene_Symbol", "Gene_Name", "gene_symbol", "gene_name"),
  mark_purity = T,
  remove_low_purity_sample = T,
  low_purity_threshold = 0.3,
  mark_lowexpressgene_pct = T,
  lowexpression_threshold = 1,
  remove_sample_with_intense_lowexpressgene = T,
  outlier_lowexpressgene_pct_factor = 1.5,
  remove_lowexpressgene = T,
  sample_frequency_threshold = 0.5
)
}
\arguments{
\item{RNA_sample_info}{\code{data.frame()}.  RNA sample information table containing at least a sample id column, along with other information.}

\item{sample_id_col}{\code{character()}. Column name for sample_id column in RNA_sample info. Default set as "Sample_ID".}

\item{expressions}{\code{data.frame()}. Gene expression table.}

\item{gene_id_col}{\code{character()}. Use row names as gene id, otherwise, provide a column name containing gene id. Default set as "rowname".}

\item{mark_library_size}{\code{logical(1)}. Mark RNA sample information table with sample library size. Default set as TRUE.}

\item{adjusted_by_library_size}{\code{logical(1)}. Normalize expressions with corresponding sample library size. Default set as TRUE.}

\item{tolerant_library_size_factor}{\code{numeric()}. When the ratio of maximal to minimal library size exceeds this threshold (default set as 1.1), trigger normalization expression by sample library size.}

\item{protein_coding_ensemble2symbol_table}{\code{data.frame()}.  Protein-coding gene table containing ensemble gene id and hugo symbol, used for extraction of protein coding expressions and convert ensemble gene id to hugo symbol}

\item{ens_id_col}{\code{character()}. Column name for ensemble gene id column in protein coding ensemble2symbol table. Default set as "Gene_ID".}

\item{symbol_col}{\code{character()}. Column name for hugo symbol column in protein coding ensemble2symbol table. Default set as "Gene_Symbol".}

\item{mark_purity}{\code{logical(1)}. Mark RNA sample information table with sample purity from estimate algorithm. Default set as TRUE.}

\item{remove_low_purity_sample}{\code{logical(1)}. Remove sample with tumor purity lower than low_purity_threshold. Default set as TRUE}

\item{low_purity_threshold}{\code{numeric()}. If remove_low_purity_sample as TRUE, sample with purity lower than low_purity_threshold (Default set at 0.3) will be excluded.}

\item{mark_lowexpressgene_pct}{\code{logical(1)}.Mark RNA sample information table with sample low expressed gene percentage. Default set as TRUE.}

\item{lowexpression_threshold}{\code{numeric()}. Threshold used to define whether gene is low expressed. Default set at 1.}

\item{remove_sample_with_intense_lowexpressgene}{\code{logical(1)}. Remove sample with high percentage of low expressed gene. Default set as TRUE.}

\item{outlier_lowexpressgene_pct_factor}{\code{numeric()}. If remove_sample_with_intense_lowexpressgene is TRUE, sample with low expressed gene percentage over the product of outlier_lowexpressgene_pct_factor (default set at 1.5) and average low expressed gene percentage across all samples will be excluded.}

\item{remove_lowexpressgene}{\code{logical(1)}.  Remove low expressed genes. Default set as TRUE.}

\item{sample_frequency_threshold}{\code{numeric()}. If remove_lowexpressgene is TRUE, genes with low expression rate (out of all samples) greater or equal to sample_frequency_threshold (default=0.5) will be removed.}
}
\value{
\code{list()}, contains marked original RNA sample information table, filtered clean RNA sample information table and filtered clean and consistent protein expressions.
}
\description{
Mark RNA sample information table with is_sequenced information, sample library size,  sample tumorpurity,sample low expressed gene percentage, and extract protein coding genes, and finally remove low express gene and prepare clean RNA sample information table and consistent protein coding mRNA expressions.
}
\examples{
results<-prepare_clean_RNA_sample_info_and_protein_expressions(RNA_sample_info=RNA_sample_info,sample_id_col = "Sample_ID",expressions = expressions,gene_id_col = "gene_id",                                     #
                                                               mark_library_size = T,adjusted_by_library_size = T,tolerant_library_size_factor = 1.1,                                                             #
                                                               protein_coding_ensemble2symbol_table = protein_coding_ensemble2symbol,ens_id_col = "Gene_ID",symbol_col = "Gene_Symbol",                           #
                                                               mark_purity = T,remove_low_purity_sample = T,low_purity_threshold = 0.3,                                                                           #
                                                               mark_lowexpressgene_pct = T,lowexpression_threshold = 1,remove_sample_with_intense_lowexpressgene = T,outlier_lowexpressgene_pct_factor = 1.5,     #
                                                               remove_lowexpressgene = T,sample_frequency_threshold = 0.5)

}
