% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_unsupervised_data.R
\name{prepare_unsupervised_data}
\alias{prepare_unsupervised_data}
\title{prepare_unsupervised_data}
\usage{
prepare_unsupervised_data(
  expressions,
  method = c("MAD", "CV", "DQ", "GUMBEL"),
  mad_top_n = -1,
  mad_top_quantile = 0.75,
  cv_top_n = -1,
  cv_top_mean_quantile = 0.5,
  dq_top_mean_quantile = 0.5,
  dq_top_var_quantile = 0.5,
  gumbel_p_cutoff = 0.1,
  remove_outlier = F
)
}
\arguments{
\item{expressions}{\code{data.frame()}. Gene expressions table, optimally clean and filtered by prepare_clean_RNA_sample_info_and_protein_expressions.}

\item{method}{\code{character()}. Methods for calculation of variance, and etc, and later used by filtering. CV for coefficient of variance, DQ for quantile of both mean and variance (dual quantile), GUMBLE for gumbel distribution of mad. Default set as "MAD".}

\item{mad_top_n}{\code{numeric()}. Number of genes to be kept with top variance, if it is -1 keep all genes. Default set as -1.}

\item{mad_top_quantile}{\code{numeric()}. Percentage of genes to be kept with top mad values if mad_top_n is -1. Default set as 0.75.}

\item{cv_top_n}{\code{numeric()}. Number of genes to be kept with top cv values, if it is -1 keep all genes. Default set as -1,}

\item{cv_top_mean_quantile}{\code{numeric()}. Percentage of genes to be kept with top mad values if cv_top_n is -1. Default set as 0.5.}

\item{dq_top_mean_quantile}{\code{numeric()}. Genes to be kept with mean value quantile above defined dq_top_mean_quantile. Default set as 0.5.}

\item{dq_top_var_quantile}{\code{numeric()}. Genes to be kept with variance value quantile above defined dq_top_var_quantile. Default set as 0.5.}

\item{gumbel_p_cutoff}{\code{numeric()}. Genes with p value of MAD gumbel distribution less than gumbel_p_cutoff will be kept. Default set as 0.1.}

\item{remove_outlier}{\code{logical(1)}. Whether to remove outlier value before calculation of variance, and etc. Default set as FALSE.}
}
\value{
a \code{data.frame()} of a subset of expressions.
}
\description{
extract qualified subset of data with most variance for unsupervised analysis
}
\examples{
results<-prepare_unsupervised_data(log2_expression,method="MAD",mad_top_n=1000,remove_outlier=F)
}
