% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runFromBasedir}
\alias{runFromBasedir}
\title{Get run id from directory}
\usage{
runFromBasedir()
}
\value{
the run id
}
\description{
This function returns the run id
by parsign the name of the directory
in which the file/s is/are
}
