% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runFromHeader}
\alias{runFromHeader}
\title{Get run id from sequence header}
\usage{
runFromHeader(nrecords = 1)
}
\arguments{
\item{nrecords}{numeric. The number of record that must be read
before returning the run id}

\item{path}{character vector. The paths to the sequence file/s}
}
\value{
the run id as character
}
\description{
This function is used internally by other
constructors to detect the run id of each file.
The id is parsed from the sequence header.
}
