% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{runTask}
\alias{runTask}
\title{Run a Task on the experiment}
\usage{
runTask(obj, task, ..., check.out = F, quiet = T)
}
\arguments{
\item{obj}{a \link{Task} object}

\item{task}{a function. The first argument of
this function will be the \link{Experiment}
object contained in the \link{Task}}

\item{...}{additianl arguments to be passed
to the funciton call}

\item{check.out}{if \code{TRUE} the \link{Task} won't be
executedon existing output file/s}

\item{quiet}{if \code{TRUE} no message will be
displayed}
}
\value{
a \link{Task} object
}
\description{
a Task (namely a function) will be
called using the experiment as first
input.
}
