% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setOut}
\alias{setOut}
\title{Format function to be used with
\link{runTask} method. This
function is a wrapper for the
\link{setOutput} method}
\usage{
setOut(value)
}
\arguments{
\item{value}{either a
character vector (for single end experiments)
or a character matrix with two columns
(for paired end samples). In alternative
it can be a function that will be used for
building output names.}
}
\value{
a TargetedExperiment
}
\description{
Format function to be used with
\link{runTask} method. This
function is a wrapper for the
\link{setOutput} method
}
