% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SingleEndSamples-class}
\alias{SingleEndSamples-class}
\alias{.SingleEndSamples}
\title{Single end experiment class}
\value{
a single end experiment
}
\description{
This class represent a standard single-end experiment.
In addition to the \link[=.TargetedExperimet]{TargetedExperimet}
class it provides a vector of file paths and a list of
output paths. They can both be used in the context
of other methods such as \code{\link[=by_sample]{by_sample()}}, for
sample wise computation and \code{\link[=by_run]{by_run()}} for run wise
computation.
}
\details{
Instances of this class should be created using the
\code{\link[=getSingleEndFromFilename]{getSingleEndFromFilename()}} function.
}
\section{Slots}{

\describe{
\item{\code{files}}{the paths to sequence files}

\item{\code{output}}{the paths to output file}
}}

\seealso{
\code{\link[=getSingleEndFromFilename]{getSingleEndFromFilename()}}, for class
creation and \code{\link[=by_sample]{by_sample()}} and \code{\link[=by_run]{by_run()}} for
computation
}
