% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{splitBy}
\alias{splitBy}
\alias{splitByFile}
\alias{splitBySample}
\alias{splitByRun}
\alias{splitByOutput}
\title{Split an Experiment into chunks}
\usage{
splitByFile(obj)

splitBySample(obj)

splitByRun(obj)

splitByOutput(obj)
}
\arguments{
\item{obj}{Experiment, the experiment to split}
}
\value{
a list of experiments splitted
according to the method used
}
\description{
\code{splitByFile} splits an Experiment into
one chunk for each input file

\code{splitBySample} splits an Experiment into
one chunk for each sample

\code{splitByRun} splits an Experiment into
one chunk for each run

\code{splitByOutput} splits an Experiment into
one chunk for each output provided. If the
Experiment is a paired-end experiment, both
forward and reverse output will be merged
and the Experiment will be spitted accordingly
}
