% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sysByAll}
\alias{sysByAll}
\title{Run a system call on the whole experiment}
\usage{
sysByAll(obj, cmd, ..., stdout = "", stderr = "", thread = 1)
}
\arguments{
\item{obj}{the experiment}

\item{cmd}{the system command}

\item{...}{argument/s to the command}

\item{stdout}{passed to stdout of \link{system2}}

\item{stderr}{passed to stderr of \link{system2}}

\item{thread}{the number of parallel process to execute}
}
\value{
an object
}
\description{
This function will execute a system call
for the whole experiment.
If threads > 1 and the doParallel package
has been installed, multiple processes will
be executed at the same time.
}
