% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sysByFct}
\alias{sysByFct}
\title{Launches a system call by splitting the experiment
based on a factor}
\usage{
sysByFct(obj, cmd, by, ..., stdout = "", stderr = "", thread = 1)
}
\arguments{
\item{obj}{the experiment}

\item{cmd}{the system command}

\item{by}{a factor. The experiment will be spitted according
to this factor}

\item{...}{argument/s to the command}

\item{stdout}{passed to stdout of \link{system2}}

\item{stderr}{passed to stderr of \link{system2}}

\item{thread}{the number of parallel process to execute}
}
\value{
a list of object
}
\description{
This function will execute a system call
for each set of the experiment.
If threads > 1 and the doParallel package
has been installed, multiple processes will
be executed at the same time.
}
