% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sysByMethods}
\alias{sysByMethods}
\alias{sysBySample}
\alias{sysByRun}
\alias{sysByFct}
\alias{sysByAll}
\alias{sysByFile}
\title{Set of methods to apply a system call
over an Experiment object}
\usage{
sysBySample(obj, cmd, ..., stdout = "", stderr = "", thread = 1)

sysByRun(obj, cmd, ..., stdout = "", stderr = "", thread = 1)

sysByFct(obj, cmd, by, ..., stdout = "", stderr = "", thread = 1)

sysByAll(obj, cmd, ..., stdout = "", stderr = "", thread = 1)

sysByFile(obj, cmd, ..., stdout = "", stderr = "", thread = 1)
}
\arguments{
\item{obj}{\link{Experiment}}

\item{cmd}{character, a system command}

\item{...}{argument/s to the command}

\item{stdout}{passed to stdout of \link{system2}}

\item{stderr}{passed to stderr of \link{system2}}

\item{thread}{integer, the number of parallel process to execute}

\item{by}{factor, the experiment will be spitted according
to this factor}
}
\value{
a list of objects returned by the function
}
\description{
If threads > 1 and the \link{doMC} package
has been installed, multiple processes will
be executed at the same time.

\code{sysBySample} the function is called splitting
the experiment into sample/s.

\code{sysByRun} the function is called splitting
the experiment into run/s.

\code{sysByFct} the function is called splitting
the experiment according to the factor provided.

\code{sysByAll} the function is called using all files
as input/output.

\code{sysByFile} the function is called on each
file separately.
}
