% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sysByRun}
\alias{sysByRun}
\title{Launches a system call for each run}
\usage{
sysByRun(obj, cmd, ..., stdout = "", stderr = "", thread = 1)
}
\arguments{
\item{obj}{the experiment}

\item{cmd}{the system command}

\item{...}{argument/s to the command}

\item{stdout}{passed to stdout of \link{system2}}

\item{stderr}{passed to stderr of \link{system2}}

\item{thread}{the number of parallel process to execute}
}
\value{
a list of object
}
\description{
This function will execute a system call
for each run of the experiment provided.
If threads > 1 and the doParallel package
has been installed, multiple processes will
be executed at the same time.
}
