% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{TargetedExperimet-class}
\alias{TargetedExperimet-class}
\alias{.TargetedExperimet}
\title{Main targeted experiment class}
\value{
a Targeted experiment
}
\description{
This class represents a sequencing experiment usually
involving sequencing a PRC products (such as 16S rRNA
gene). Each sample may have one or two files (single-end
or paired-end sequencing) and a run identifier.
}
\section{Slots}{

\describe{
\item{\code{samples}}{Character vector, sample names}

\item{\code{run}}{Character vector, the run id for each sample}

\item{\code{step}}{Character, the step of analysis}

\item{\code{n}}{Numeric vector, the number of samples}
}}

