% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Task-class}
\alias{Task-class}
\alias{.Task}
\title{A task object}
\value{
A Task object
}
\description{
An object that represents a possible
task for an experiment.
}
\section{Slots}{

\describe{
\item{\code{task}}{function. The real task. The first argument
must be a \link{TargetedExperiment} object}

\item{\code{exp}}{TargetedExperiment. The experiment}

\item{\code{out}}{ANY. Output of the task will be stored here}
}}

