% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Task}
\alias{Task}
\title{Constructor for Task objects}
\usage{
Task(exp, task)
}
\arguments{
\item{exp}{TargetedExperiment. The experiment}

\item{task}{function. The real task}
}
\value{
a Task object
}
\description{
Constructor for Task objects
}
