% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{toSingleEnd}
\alias{toSingleEnd}
\title{Transform a paired-end object into two single-end
objects}
\usage{
toSingleEnd(obj)
}
\arguments{
\item{obj}{PairedSamples, the experiment to be
transformed}
}
\value{
list, a list containing two SingleEndSamples
}
\description{
This method transforms a paired-end object into
two single-end objects, one for forward file/s
and the other for reverse file/s
}
