% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_analysis.R
\name{unsupervised_analysis}
\alias{unsupervised_analysis}
\title{unsupervised analysis}
\usage{
unsupervised_analysis(
  dat,
  labels,
  cols,
  scale = c("row", "column", "none"),
  run_umap = T,
  umap_params = NULL,
  umap_config = umap.defaults,
  run_tsne = T,
  tsne_params = NULL,
  run_pca = T,
  pca_params = NULL,
  run_mds = T,
  mds_params = NULL,
  run_heatmap = T,
  heatmap_params = NULL
)
}
\arguments{
\item{dat}{\code{data.frame()}. Gene expressions table.}

\item{labels}{\code{vector(mode = "character")}. Sample labels corresponding to all samples (column names) of \code{dat}.}

\item{cols}{\code{list()}. Colors to control presentation of labels.}

\item{scale}{\code{character(1)}. Direction to scale \code{dat}. Use "row" to scale by row; "column" to scale by column; "none" for no scaling. Default set as "row".}

\item{run_umap}{\code{logical(1)}. Whether to run UMAP analysis. Default set as TRUE.}

\item{umap_params}{\code{list()}. Params passed to UMAP. Default is NULL.}

\item{umap_config}{\code{list()}. Extra config param passed to UMAP. Default is umap.defaults.}

\item{run_tsne}{\code{logical(1)}. Whether to run TSNE analysis. Default set as TRUE.}

\item{tsne_params}{\code{list()}. Params passed to TSNE. Default is NULL.}

\item{run_pca}{\code{logical(1)}. Whether to run PCA analysis. Default set as TRUE.}

\item{pca_params}{\code{list()}. Params passed to PCA. Default is NULL.}

\item{run_mds}{\code{logical(1)}. Whether to run MDS analysis. Default set as TRUE.}

\item{mds_params}{\code{list()}. Params passed to MDS. Default is NULL.}

\item{run_heatmap}{\code{logical(1)}. Whether to run clustering analysis with heatmap. Default set as TRUE.}

\item{heatmap_params}{list, default NULL, params passed to Heatmap}
}
\value{
\code{list()} containing results from UMPA, TSNE, PCA, MDS, Cluster(Heatmap) analysis.
}
\description{
MAP, TSNE, PCA, MDS and unsupervised clustering of expressions
}
\examples{
\dontrun{
results<-unsupervised_analysis(unsupervised_data,labels=clean_RNA_sample_info$PROTOCOL)
# export figure
tiff(filename ="unsupervised_cluster.tiff",width = 12,height=8,units = "in",res=300,compression = "lzw")
grid.arrange(arrangeGrob(
  results$plots$umap,
  results$plots$tsne,
  results$plots$pca$`2D`,
  results$plots$mds,
  results$plots$heatmap,
  layout_matrix = matrix(c(1:4,rep(5,4)),ncol = 2,byrow = T)),
  cowplot::get_legend(results$plots$umap+theme(legend.position = "top")),
  nrow = 2, heights = c(12, 1))
dev.off()
}

}
