% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validExperimentObject}
\alias{validExperimentObject}
\title{Validation function for \link{Experiment}}
\usage{
validExperimentObject(object)
}
\arguments{
\item{object}{a \link{Experiment}}
}
\value{
logical. TRUE if the object is valid
}
\description{
Validation function for \link{Experiment}
}
