% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validPairedSamplesObject}
\alias{validPairedSamplesObject}
\title{Validation function for \link{PairedSamples}}
\usage{
validPairedSamplesObject(object)
}
\arguments{
\item{object}{a \link{PairedSamples}}
}
\value{
logical. TRUE if the object is valid
}
\description{
Validation function for \link{PairedSamples}
}
