% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validSingleEndSamplesObject}
\alias{validSingleEndSamplesObject}
\title{Validation function for \link{SingleEndSamples}}
\usage{
validSingleEndSamplesObject(object)
}
\arguments{
\item{object}{a \link{SingleEndSamples}}
}
\value{
logical. TRUE if the object is valid
}
\description{
Validation function for \link{SingleEndSamples}
}
