% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validTargetedExperimetObject}
\alias{validTargetedExperimetObject}
\title{Validation function for \link{TergetedExperiment}}
\usage{
validTargetedExperimetObject(object)
}
\arguments{
\item{object}{a \link{TergetedExperiment}}
}
\value{
logical. TRUE if the object is valid
}
\description{
Validation function for \link{TergetedExperiment}
}
