% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{accessors}
\alias{accessors}
\alias{step}
\alias{step<-}
\alias{N}
\alias{files}
\alias{output}
\alias{samples}
\alias{run}
\title{Set of methods to access or to set
slots of a TargetedExperiment}
\usage{
step(obj)

step(obj) <- value

N(obj)

files(obj)

output(obj)

samples(obj)

run(obj)
}
\arguments{
\item{obj}{a TargetedExperiment object}

\item{value}{the value to set}
}
\description{
\code{step} returns/set the step of analysis
from a \link{TargetedExperiment} object.

\code{N} returns the number of samples.

\code{files} and \code{output} return the input/output
files from a \link{TargetedExperiment} object. If object
is a \link{SingleEndSamples} it will return a
vector of files otherwise it will return
a data frame with forward and reverse files.

\code{samples} returns the samples from
a \link{TargetedExperiment} object.

\code{run} returns the samples from
a \link{TargetedExperiment} object.
}
