% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{accessors}
\alias{accessors}
\alias{step}
\alias{step<-}
\alias{setStep}
\alias{N}
\alias{files}
\alias{output}
\alias{setOutput}
\alias{samples}
\alias{run}
\alias{nrun}
\alias{output.exists}
\alias{input.exists}
\alias{up2date}
\title{Set of methods to access or to set
slots of a Experiment}
\usage{
step(obj)

step(obj) <- value

setStep(obj, value)

N(obj)

files(obj)

output(obj)

setOutput(obj, value)

samples(obj)

run(obj)

nrun(obj)

output.exists(obj)

input.exists(obj)

up2date(obj)
}
\arguments{
\item{obj}{\link{Experiment}}

\item{value}{the value to set. For the
\code{setOutput} function, value must be a
character vector (or a matrix for paired samples)
or a function}

\item{obj}{\link{Experiment}}

\item{obj}{\link{Experiment}}

\item{obj}{\link{Experiment}}
}
\description{
\code{step} returns/set the step of analysis
from a \link{Experiment} object.

\code{setStep} sets the step of analysis
of an \link{Experiment} object.

\code{N} returns the number of samples.

\code{files} and \code{output} return the input/output
files from a \link{Experiment} object. If object
is a \link{SingleEndSamples} it will return a
vector of files otherwise it will return
a data frame with forward and reverse files.

\code{setOutput} is used ot change tho output of
a given experiment. The \code{value} parameter must be a
character vector (for single end experiments)
or a character matrix with two columns
(for paired end samples). In alternative
it can be a function that will be used for
building output names.

\code{samples} returns the samples from
a \link{Experiment} object.

\code{run} returns the samples from
a \link{Experiment} object.

\code{nrun} returns the number of run from
a \link{Experiment} object.

\code{output.exists} returns \code{TRUE}
for each output file that is found
on the hard drive. If the experiment
is a paired-end experiment,
the method will return \code{TRUE} only
if both forward and reverse
output files exist.

\code{input.exists} returns \code{TRUE}
for each input file that is found
on the hard drive. If the experiment
is a paired-end experiment,
the method will return \code{TRUE} only
if both forward and reverse
input files exist.

\code{up2date} returns \code{TRUE}
for each input and output file up to date.
Namely, the timestamp of each input file
must be smaller than its output file.
}
