% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{runTask}
\alias{runTask}
\title{Run a task}
\usage{
runTask(obj, out.exp = function(obj) {     obj }, force = F,
  quiet = T)
}
\arguments{
\item{obj}{the task}

\item{out.exp}{function. This function will be called at the
end of the task to transform the experiment}

\item{force}{if \code{TRUE} output will be overridden
even if it exists.}

\item{quiet}{if \code{FALSE} no messages will be displayed}
}
\value{
a task object with after running the task
}
\description{
Run a task
}
