% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Task-class}
\alias{Task-class}
\alias{Task}
\title{A Task object}
\value{
A Task object
}
\description{
An object that represents a Task of
an experiment.
}
\section{Slots}{

\describe{
\item{\code{exp}}{\link{Experiment}, the experiment}

\item{\code{out}}{list, output of the task will be stored here}
}}

