% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{task.accessors}
\alias{task.accessors}
\alias{getExp}
\alias{getOut}
\title{Set of methods to access the
slots of a \link{Task} object}
\usage{
getExp(obj)

getOut(obj)
}
\arguments{
\item{obj}{a Task object}

\item{obj}{Task.}

\item{obj}{}
}
\value{
the Experiment

the output from a Task object
}
\description{
\code{getExp} returns the experiment from
a \link{Task} object

\code{getOut} returns the output
from a task object
}
